Move Traits to their own folder
